hardenend parser combinator library

+++ alef/src/main/scala/beklash/alef/Model.scala
@@ -1,7 +1,7 @@
 package beklash
 package alef
 
-type Value  = Int
+type Value  = Double
 type Name   = String
 type Input  = Map[Name, Value]
 type Error  = String
